Removed duplicate implementation of Query and ResultNamingStrategy to fix #18.
Some cleanup done to ResultNamingStrategy (more to come, there are still a few parts I'm not happy with). It took me forever to realize that ResultNamingStrategy was mainly a template engine (which has been extracted).
