+++ jmxtrans2-core/pom.xml
@@ -38,13 +38,9 @@
         is what you need.</description>
 
     <properties>
-        <verify.bundleLineRate>0.78</verify.bundleLineRate>
-        <verify.classBranchRate>0.0</verify.classBranchRate>
         <verify.classLineRate>0.40</verify.classLineRate>
-        <verify.coverage.threshold>77</verify.coverage.threshold>
-        <verify.mutation.threshold>54</verify.mutation.threshold>
-        <verify.packageBranchRate>0.26</verify.packageBranchRate>
-        <verify.packageLineRate>0.47</verify.packageLineRate>
+        <verify.mutation.threshold>56</verify.mutation.threshold>
+        <verify.packageBranchRate>0.49</verify.packageBranchRate>
     </properties>
 
     <dependencies>
@@ -128,7 +124,7 @@
                         <!-- no need to test generated classes -->
                         <excludedClass>org.jmxtrans.core.config.jaxb.*</excludedClass>
                         <!-- some problem on travis with this test, let's disable it until we understand better -->
-                        <excludedClass>org.jmxtrans.core.query.embedded.QueryTest</excludedClass>
+                        <excludedClass>org.jmxtrans.core.query.QueryTest</excludedClass>
                     </excludedClasses>
                 </configuration>
             </plugin>
