+++ jmxtrans2-core/src/main/java/org/jmxtrans/core/config/JmxTransBuilder.java
@@ -46,8 +46,8 @@ import org.jmxtrans.core.log.Logger;
 import org.jmxtrans.core.log.LoggerFactory;
 import org.jmxtrans.core.monitoring.MBeanRegistry;
 import org.jmxtrans.core.monitoring.SelfNamedMBean;
-import org.jmxtrans.core.query.embedded.ResultNameStrategy;
-import org.jmxtrans.core.query.embedded.Server;
+import org.jmxtrans.core.query.ResultNameStrategy;
+import org.jmxtrans.core.query.Server;
 import org.jmxtrans.core.scheduler.JmxTransThreadFactory;
 import org.jmxtrans.core.scheduler.NaiveScheduler;
 import org.jmxtrans.core.scheduler.QueryGenerator;
