+++ jmxtrans2-core/src/main/java/org/jmxtrans/core/config/StandardConfiguration.java
@@ -29,7 +29,7 @@ import javax.annotation.concurrent.ThreadSafe;
 
 import org.jmxtrans.core.output.OutputWriter;
 import org.jmxtrans.core.query.Invocation;
-import org.jmxtrans.core.query.embedded.Server;
+import org.jmxtrans.core.query.Server;
 import org.jmxtrans.utils.time.Interval;
 
 @ThreadSafe // TODO: synchronization is overly aggressive
