+++ jmxtrans2-core/src/main/java/org/jmxtrans/core/config/XmlConfigParser.java
@@ -59,11 +59,11 @@ import org.jmxtrans.core.log.LoggerFactory;
 import org.jmxtrans.core.output.MetricCollectingOutputWriter;
 import org.jmxtrans.core.output.OutputWriter;
 import org.jmxtrans.core.output.OutputWriterFactory;
+import org.jmxtrans.core.query.InProcessServer;
 import org.jmxtrans.core.query.Invocation;
-import org.jmxtrans.core.query.embedded.InProcessServer;
-import org.jmxtrans.core.query.embedded.Query;
-import org.jmxtrans.core.query.embedded.QueryAttribute;
-import org.jmxtrans.core.query.embedded.RemoteServer;
+import org.jmxtrans.core.query.Query;
+import org.jmxtrans.core.query.QueryAttribute;
+import org.jmxtrans.core.query.RemoteServer;
 import org.jmxtrans.utils.io.Resource;
 import org.jmxtrans.utils.io.StandardResource;
 import org.jmxtrans.utils.time.Clock;
