--- jmxtrans2-core/src/main/java/org/jmxtrans/core/naming/ExpressionLanguageEngineImpl.java
@@ -1,210 +1,0 @@
-/**
- * The MIT License
- * Copyright (c) 2014 JMXTrans Team
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-package org.jmxtrans.core.naming;
-
-import java.net.InetAddress;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.concurrent.Callable;
-
-import javax.annotation.Nonnull;
-import javax.management.ObjectName;
-
-import org.jmxtrans.core.log.Logger;
-import org.jmxtrans.core.log.LoggerFactory;
-import org.jmxtrans.utils.StringUtils2;
-
-/**
- * @author <a href="mailto:cleclerc@cloudbees.com">Cyrille Le Clerc</a>
- */
-public class ExpressionLanguageEngineImpl implements ExpressionLanguageEngine {
-    protected final Logger logger = LoggerFactory.getLogger(getClass().getName());
-
-    public ExpressionLanguageEngineImpl() {
-        try {
-            InetAddress localHost = InetAddress.getLocalHost();
-            String hostName = localHost.getHostName();
-            String reversedHostName = StringUtils2.reverseTokens(hostName, ".");
-            String canonicalHostName = localHost.getCanonicalHostName();
-            String reversedCanonicalHostName = StringUtils2.reverseTokens(canonicalHostName, ".");
-            String hostAddress = localHost.getHostAddress();
-
-            registerExpressionEvaluator("hostname", hostName);
-            registerExpressionEvaluator("reversed_hostname", reversedHostName);
-            registerExpressionEvaluator("escaped_hostname", hostName.replaceAll("\\.", "_"));
-            registerExpressionEvaluator("canonical_hostname", canonicalHostName);
-            registerExpressionEvaluator("reversed_canonical_hostname", reversedCanonicalHostName);
-            registerExpressionEvaluator("escaped_canonical_hostname", canonicalHostName.replaceAll("\\.", "_"));
-            registerExpressionEvaluator("hostaddress", hostAddress);
-            registerExpressionEvaluator("escaped_hostaddress", hostAddress.replaceAll("\\.", "_"));
-        } catch (Exception e) {
-            logger.warn("Exception resolving localhost, expressions like #hostname#, #canonical_hostname# or #hostaddress# will not be available", e);
-        }
-    }
-
-    /**
-     * Function based evaluators for expressions like '#hostname#' or '#hostname_canonical#'
-     */
-    @Nonnull
-    private Map<String, Callable<String>> expressionEvaluators = new HashMap<String, Callable<String>>();
-
-    /**
-     * Replace all the '#' based keywords (e.g. <code>#hostname#</code>) by their value.
-     *
-     * @param expression the expression to resolve (e.g. <code>"servers.#hostname#."</code>)
-     * @return the resolved expression (e.g. <code>"servers.tomcat5"</code>)
-     */
-    @Nonnull
-    public String resolveExpression(@Nonnull String expression) {
-        StringBuilder result = new StringBuilder(expression.length());
-
-        int position = 0;
-        while (position < expression.length()) {
-            char c = expression.charAt(position);
-            if (c == '#') {
-                int beginningSeparatorPosition = position;
-                int endingSeparatorPosition = expression.indexOf('#', beginningSeparatorPosition + 1);
-                if (endingSeparatorPosition == -1) {
-                    throw new IllegalStateException("Invalid expression '" + expression + "', no ending '#' after beginning '#' at position " + beginningSeparatorPosition);
-                }
-                String key = expression.substring(beginningSeparatorPosition + 1, endingSeparatorPosition);
-                Callable<String> expressionProcessor = expressionEvaluators.get(key);
-                String value;
-                if (expressionProcessor == null) {
-                    value = "#unsupported_expression#";
-                    logger.info("Unsupported expression '" + key + "'");
-                } else {
-                    try {
-                        value = expressionProcessor.call();
-                    } catch (Exception e) {
-                        value = "#expression_error#";
-                        logger.warn("Error evaluating expression '" + key + "'", e);
-                    }
-                }
-                StringUtils2.appendEscapedNonAlphaNumericChars(value, false, result);
-                position = endingSeparatorPosition + 1;
-
-            } else {
-                result.append(c);
-                position++;
-            }
-        }
-        if (logger.isDebugEnabled()) logger.debug("resolveExpression(" + expression + "): " + result);
-        return result.toString();
-
-    }
-
-    @Nonnull
-    public String resolveExpression(@Nonnull String expression, @Nonnull ObjectName exactObjectName) {
-
-        StringBuilder result = new StringBuilder(expression.length());
-
-        int position = 0;
-        while (position < expression.length()) {
-            char c = expression.charAt(position);
-            if (c == '%') {
-                int beginningSeparatorPosition = position;
-                int endingSeparatorPosition = expression.indexOf('%', beginningSeparatorPosition + 1);
-                if (endingSeparatorPosition == -1) {
-                    throw new IllegalStateException("Invalid expression '" + expression + "', no ending '%' after beginning '%' at position " + beginningSeparatorPosition);
-                }
-                String key = expression.substring(beginningSeparatorPosition + 1, endingSeparatorPosition);
-                String value = exactObjectName.getKeyProperty(key);
-                if (value == null) {
-                    value = "null";
-                }
-                StringUtils2.appendEscapedNonAlphaNumericChars(value, result);
-                position = endingSeparatorPosition + 1;
-            } else if (c == '#') {
-                int beginningSeparatorPosition = position;
-                int endingSeparatorPosition = expression.indexOf('#', beginningSeparatorPosition + 1);
-                if (endingSeparatorPosition == -1) {
-                    throw new IllegalStateException("Invalid expression '" + expression + "', no ending '#' after beginning '#' at position " + beginningSeparatorPosition);
-                }
-                String key = expression.substring(beginningSeparatorPosition + 1, endingSeparatorPosition);
-                Callable<String> expressionProcessor = expressionEvaluators.get(key);
-                String value;
-                if (expressionProcessor == null) {
-                    value = "#unsupported_expression#";
-                    logger.info("Unsupported expression '" + key + "'");
-                } else {
-                    try {
-                        value = expressionProcessor.call();
-                    } catch (Exception e) {
-                        value = "#expression_error#";
-                        logger.warn("Error evaluating expression '" + key + "'", e);
-                    }
-                }
-                StringUtils2.appendEscapedNonAlphaNumericChars(value, false, result);
-                position = endingSeparatorPosition + 1;
-
-            } else {
-                result.append(c);
-                position++;
-            }
-        }
-        if (logger.isDebugEnabled())
-            logger.debug("resolveExpression(" + expression + ", " + exactObjectName + "): " + result);
-
-        return result.toString();
-    }
-
-    /**
-     * Registers an expression evaluator with a static value.
-     */
-    public void registerExpressionEvaluator(String expression, Callable<String> evaluator) {
-        expressionEvaluators.put(expression, evaluator);
-    }
-
-    /**
-     * Registers an expression evaluator with a static value.
-     */
-    public void registerExpressionEvaluator(String expression, String value) {
-        expressionEvaluators.put(expression, new StaticEvaluator(value));
-    }
-
-    @Nonnull
-    public Map<String, Callable<String>> getExpressionEvaluators() {
-        return expressionEvaluators;
-    }
-
-    public static class StaticEvaluator implements Callable<String> {
-        final String value;
-
-        public StaticEvaluator(String value) {
-            this.value = value;
-        }
-
-        @Override
-        public String call() throws Exception {
-            return value;
-        }
-
-        @Override
-        public String toString() {
-            return "StaticStringCallable{" +
-                    "value='" + value + '\'' +
-                    '}';
-        }
-    }
-}
