--- jmxtrans2-core/src/main/java/org/jmxtrans/core/naming/ResultNameStrategyImpl.java
@@ -1,152 +1,0 @@
-/**
- * The MIT License
- * Copyright (c) 2014 JMXTrans Team
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-package org.jmxtrans.core.naming;
-
-import java.util.Collections;
-import java.util.Iterator;
-import java.util.List;
-
-import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
-import javax.management.ObjectName;
-
-import org.jmxtrans.core.log.Logger;
-import org.jmxtrans.core.log.LoggerFactory;
-import org.jmxtrans.core.query.Query;
-import org.jmxtrans.core.query.ResultNameStrategy;
-import org.jmxtrans.utils.StringUtils2;
-
-/**
- * Build a {@linkplain org.jmxtrans.core.results.QueryResult#name} from a collected metric ({@linkplain org.jmxtrans.core.query.Query}).
- * <p/>
- * Build name must be escaped to be compatible with all OutputWriters.
- * The approach is to escape non alpha-numeric chars.
- * <p/>
- * Expressions support '#' based keywords (e.g. <code>#hostname#</code>) and with '%' based variables mapped to objectname properties.
- * <p/>
- * Supported '#' based 'functions':
- * <table>
- * <tr>
- * <th>Function</th>
- * <th>Description</th>
- * <th>Sample</th>
- * </tr>
- * <tr>
- * <th><code>#hostname#</code></th>
- * <td>localhost - hostname {@link java.net.InetAddress#getHostName()}</td>
- * <td></td>
- * </tr>
- * <tr>
- * <th><code>#reversed_hostname#</code></th>
- * <td>reversed localhost - hostname {@link java.net.InetAddress#getHostName()}</td>
- * <td></td>
- * </tr>
- * <tr>
- * <th><code>#escaped_hostname#</code></th>
- * <td>localhost - hostname {@link java.net.InetAddress#getHostName()} with '.' replaced by '_'</td>
- * <td></td>
- * </tr>
- * <tr>
- * <th><code>#canonical_hostname#</code></th>
- * <td>localhost - canonical hostname {@link java.net.InetAddress#getCanonicalHostName()}</td>
- * <td><code>server1.ecommerce.mycompany.com</code></td>
- * </tr>
- * <tr>
- * <th><code>#reversed_canonical_hostname#</code></th>
- * <td>reversed localhost - canonical hostname {@link java.net.InetAddress#getCanonicalHostName()}</td>
- * <td><code>com.mycompany.ecommerce.server1</code></td>
- * </tr>
- * <tr>
- * <th><code>#escaped_canonical_hostname#</code></th>
- * <td>localhost - canonical hostname {@link java.net.InetAddress#getCanonicalHostName()} with '.' replaced by '_'</td>
- * <td><code>server1_ecommerce_mycompany_com</code></td>
- * </tr>
- * <tr>
- * <th><code>#hostaddress#</code></th>
- * <td>localhost - hostaddress {@link java.net.InetAddress#getHostAddress()}</td>
- * <td></td>
- * </tr>
- * <tr>
- * <th><code>#escaped_hostname#</code></th>
- * <td>localhost - hostaddress {@link java.net.InetAddress#getHostAddress()} with '.' replaced by '_'</td>
- * <td></td>
- * </tr>
- * </table>
- *
- * @author <a href="mailto:cleclerc@cloudbees.com">Cyrille Le Clerc</a>
- */
-public class ResultNameStrategyImpl implements ResultNameStrategy {
-
-    protected final Logger logger = LoggerFactory.getLogger(getClass().getName());
-
-    @Nonnull
-    private ExpressionLanguageEngine expressionLanguageEngine = new ExpressionLanguageEngineImpl();
-
-    @Nonnull
-    @Override
-    public String getResultName(@Nonnull Query query, @Nonnull ObjectName objectName) {
-        return getResultName(query, objectName, null);
-    }
-
-    @Nonnull
-    @Override
-    public String getResultName(@Nonnull Query query, @Nonnull ObjectName objectName, @Nullable String key) {
-        return expressionLanguageEngine.resolveExpression(query.getResultAlias(), objectName);
-    }
-
-    /**
-     * Transforms an {@linkplain javax.management.ObjectName} into a plain {@linkplain String} only composed of (a->Z, A-Z, '_').
-     * <p/>
-     * '_' is the escape char for not compliant chars.
-     */
-    @Nonnull
-    protected String escapeObjectName(@Nonnull ObjectName objectName) {
-        StringBuilder result = new StringBuilder();
-        StringUtils2.appendEscapedNonAlphaNumericChars(objectName.getDomain(), result);
-        result.append('.');
-        List<String> keys = Collections.list(objectName.getKeyPropertyList().keys());
-        Collections.sort(keys);
-        for (Iterator<String> it = keys.iterator(); it.hasNext(); ) {
-            String key = it.next();
-            StringUtils2.appendEscapedNonAlphaNumericChars(key, result);
-            result.append("__");
-            StringUtils2.appendEscapedNonAlphaNumericChars(objectName.getKeyProperty(key), result);
-            if (it.hasNext()) {
-                result.append('.');
-            }
-        }
-        if (logger.isDebugEnabled())
-            logger.debug("escapeObjectName(" + objectName + "): " + result);
-        return result.toString();
-    }
-
-    @Nonnull
-    public ExpressionLanguageEngine getExpressionLanguageEngine() {
-        return expressionLanguageEngine;
-    }
-
-    public void setExpressionLanguageEngine(@Nonnull ExpressionLanguageEngine expressionLanguageEngine) {
-        this.expressionLanguageEngine = expressionLanguageEngine;
-    }
-
-}
