+++ jmxtrans2-core/src/main/java/org/jmxtrans/core/query/QueryAttribute.java
@@ -20,7 +20,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.jmxtrans.core.query.embedded;
+package org.jmxtrans.core.query;
 
 import java.util.Arrays;
 import java.util.Collection;
