+++ jmxtrans2-core/src/main/java/org/jmxtrans/core/scheduler/QueryProcessor.java
@@ -31,9 +31,9 @@ import javax.annotation.concurrent.ThreadSafe;
 import org.jmxtrans.core.log.Logger;
 import org.jmxtrans.core.log.LoggerFactory;
 import org.jmxtrans.core.output.OutputWriter;
-import org.jmxtrans.core.query.embedded.Query;
-import org.jmxtrans.core.query.embedded.ResultNameStrategy;
-import org.jmxtrans.core.query.embedded.Server;
+import org.jmxtrans.core.query.Query;
+import org.jmxtrans.core.query.ResultNameStrategy;
+import org.jmxtrans.core.query.Server;
 import org.jmxtrans.core.results.QueryResult;
 import org.jmxtrans.utils.time.Clock;
