+++ jmxtrans2-core/src/test/java/org/jmxtrans/core/query/QueryTest.java
@@ -23,132 +23,117 @@
 package org.jmxtrans.core.query;
 
 import java.lang.management.ManagementFactory;
-import java.util.Queue;
-import java.util.concurrent.ArrayBlockingQueue;
+import java.util.Iterator;
 
-import javax.annotation.Nullable;
 import javax.management.MBeanServer;
 import javax.management.ObjectName;
 
 import org.jmxtrans.core.results.QueryResult;
 
-import com.google.common.base.Optional;
-import com.google.common.base.Predicate;
-import com.google.common.collect.FluentIterable;
-import org.junit.After;
-import org.junit.Before;
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
 import org.junit.Test;
+
+import static java.util.Arrays.asList;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 /**
- * @author <a href="mailto:cleclerc@cloudbees.com">Cyrille Le Clerc</a>
+ * @author <a href="mailto:cleclerc@xebia.fr">Cyrille Le Clerc</a>
  */
 public class QueryTest {
 
-    private static final String MOCK_MBEAN_NAME = "test:type=Mock,name=mock";
-    private DummyResultNameStrategy resultNameStrategy;
-    private MBeanServer mbeanServer;
-    private Mock mock;
+    static MBeanServer mbeanServer = ManagementFactory.getPlatformMBeanServer();
+    static ObjectName mockEdenSpacePool;
+    static ObjectName mockPermGenPool;
 
-    @Before
-    public void before() throws Exception {
-        resultNameStrategy = new DummyResultNameStrategy();
-        mock = new Mock("PS Eden Space", 87359488L);
-        mbeanServer = ManagementFactory.getPlatformMBeanServer();
-        mbeanServer.registerMBean(mock, new ObjectName(MOCK_MBEAN_NAME));
+
+    @BeforeClass
+    public static void beforeClass() throws Exception {
+        mockEdenSpacePool = new ObjectName("test:type=MemoryPool,name=PS Eden Space");
+        mbeanServer.registerMBean(new MockMemoryPool("PS Eden Space", 87359488L), mockEdenSpacePool);
+        mockPermGenPool = new ObjectName("test:type=MemoryPool,name=PS Perm Gen");
+        mbeanServer.registerMBean(new MockMemoryPool("PS Perm Gen", 87752704L), mockPermGenPool);
     }
 
-    @After
-    public void after() throws Exception {
-        mbeanServer.unregisterMBean(new ObjectName(MOCK_MBEAN_NAME));
+    @AfterClass
+    public static void afterClass() throws Exception {
+        mbeanServer.unregisterMBean(mockEdenSpacePool);
+        mbeanServer.unregisterMBean(mockPermGenPool);
     }
 
     @Test
-    public void basic_attribute_return_simple_result() throws Exception {
-        Query query = new Query(MOCK_MBEAN_NAME, "CollectionUsageThreshold", resultNameStrategy);
-        Queue<QueryResult> results = new ArrayBlockingQueue<QueryResult>(1);
-
-        query.collectMetrics(mbeanServer, results);
-
+    public void basic_jmx_attribute_return_simple_result() throws Exception {
+        Iterable<QueryResult> results = Query.builder()
+                .withObjectName("test:type=MemoryPool,name=PS Eden Space")
+                .addAttribute("CollectionUsageThreshold")
+                .build()
+                .collectMetrics(mbeanServer, new ResultNameStrategy());
         assertThat(results).hasSize(1);
-        QueryResult queryResult = results.poll();
-        assertThat(queryResult).isNotNull();
-        assertThat(queryResult.getName()).isEqualTo("CollectionUsageThreshold");
-        assertThat(queryResult.getValue()).isInstanceOf(Number.class);
+
+        QueryResult result = results.iterator().next();
+        assertThat(result.getValue()).isInstanceOf(Number.class);
     }
 
     @Test
-    public void result_name_are_processed() {
-        Query query = new Query(MOCK_MBEAN_NAME, "CollectionUsageThreshold", resultNameStrategy);
-        Queue<QueryResult> results = new ArrayBlockingQueue<QueryResult>(1);
-
-        query.collectMetrics(mbeanServer, results);
-
-        assertThat(resultNameStrategy.hasBeenCalled()).isTrue();
-    }
-
-    @Test
-    public void indexed_list_attribute_return_simple_result() throws Exception {
-        Query query = new Query(MOCK_MBEAN_NAME, "IntegerList", 1, resultNameStrategy);
-        Queue<QueryResult> results = new ArrayBlockingQueue<QueryResult>(1);
-
-        query.collectMetrics(mbeanServer, results);
-
+    public void attributesAreAvailableInNamingStrategy() throws Exception {
+        Iterable<QueryResult> results = Query.builder()
+                .withObjectName("test:type=*,name=PS Eden Space")
+                .withResultAlias("memory.%type%")
+                .addAttribute("CollectionUsageThreshold")
+                .build()
+                .collectMetrics(mbeanServer, new ResultNameStrategy());
         assertThat(results).hasSize(1);
-        QueryResult queryResult = results.poll();
-        assertThat(queryResult).isNotNull();
-        assertThat(queryResult.getName()).isEqualTo("IntegerList");
-        assertThat(queryResult.getValue()).isInstanceOf(Number.class);
+
+        QueryResult result = results.iterator().next();
+        assertThat(result.getValue()).isInstanceOf(Number.class);
+        assertThat(result.getName()).isEqualTo("memory.MemoryPool.CollectionUsageThreshold");
     }
 
     @Test
-    public void non_indexed_list_attribute_return_simple_result() throws Exception {
-        Query query = new Query(MOCK_MBEAN_NAME, "IntegerList", resultNameStrategy);
-        Queue<QueryResult> results = new ArrayBlockingQueue<QueryResult>(mock.getIntegerList().size());
+    public void test_composite_jmx_attribute() throws Exception {
+        Query query = Query.builder()
+                .withObjectName("test:type=MemoryPool,name=PS Perm Gen")
+                .addAttribute(QueryAttribute.builder("Usage")
+                        .withKeys(asList("committed", "init", "max", "used"))
+                        .build())
+                .build();
+        Iterable<QueryResult> results = query.collectMetrics(mbeanServer, new ResultNameStrategy());
+        assertThat(results).hasSize(4);
 
-        query.collectMetrics(mbeanServer, results);
+        Iterator<QueryResult> queryResultIterator = results.iterator();
 
-        for (int i = 0; i < mock.getIntegerList().size(); i++) {
-            final String name = "IntegerList_" + i;
-            Optional<QueryResult> result = FluentIterable.from(results).firstMatch(new Predicate<QueryResult>() {
-                @Override
-                public boolean apply(@Nullable QueryResult input) {
-                    return input != null && name.equals(input.getName());
-                }
-            });
+        QueryResult result1 = queryResultIterator.next();
+        assertThat(result1.getValue()).isInstanceOf(Number.class);
 
-            org.assertj.guava.api.Assertions.assertThat(result).isPresent();
-            assertThat(result.get().getValue()).isInstanceOf(Number.class);
-        }
+        QueryResult result2 = queryResultIterator.next();
+        assertThat(result2.getValue()).isInstanceOf(Number.class);
     }
 
     @Test
-    public void indexed_int_array_attribute_return_simple_result() throws Exception {
-        Query query = new Query(MOCK_MBEAN_NAME, "IntArray", 1, resultNameStrategy);
-        Queue<QueryResult> results = new ArrayBlockingQueue<QueryResult>(1);
+    public void maxResultsIsHonored() throws Exception {
+        Query query = Query.builder()
+                .withObjectName("test:type=MemoryPool,name=PS Perm Gen")
+                .withMaxResults(2)
+                .addAttribute(QueryAttribute.builder("Usage")
+                        .withKeys(asList("committed", "init", "max", "used"))
+                        .build())
+                .build();
 
-        query.collectMetrics(mbeanServer, results);
+        Iterable<QueryResult> results = query.collectMetrics(mbeanServer, new ResultNameStrategy());
 
-        assertThat(results).hasSize(1);
-        QueryResult queryResult = results.poll();
-        assertThat(queryResult).isNotNull();
-        assertThat(queryResult.getName()).isEqualTo("IntArray");
-        assertThat(queryResult.getValue()).isInstanceOf(Number.class);
+        assertThat(results).hasSize(2);
     }
 
     @Test
-    public void indexed_integer_array_attribute_return_simple_result() throws Exception {
-        Query query = new Query(MOCK_MBEAN_NAME, "IntegerArray", 1, resultNameStrategy);
-        Queue<QueryResult> results = new ArrayBlockingQueue<QueryResult>(1);
+    public void queryWithSameNameAreEquals() {
+        Query query1 = Query.builder()
+                .withObjectName("test:type=MemoryPool,name=PS Perm Gen")
+                .build();
+        Query query2 = Query.builder()
+                .withObjectName("test:type=MemoryPool,name=PS Perm Gen")
+                .build();
 
-        query.collectMetrics(mbeanServer, results);
-
-        assertThat(results).hasSize(1);
-        QueryResult queryResult = results.poll();
-        assertThat(queryResult).isNotNull();
-        assertThat(queryResult.getName()).isEqualTo("IntegerArray");
-        assertThat(queryResult.getValue()).isInstanceOf(Number.class);
+        assertThat(query1).isEqualTo(query2);
     }
-
 }
