+++ jmxtrans2-core/src/test/java/org/jmxtrans/core/scheduler/FullSchedulingTest.java
@@ -36,10 +36,10 @@ import javax.management.MBeanServerConnection;
 
 import org.jmxtrans.core.lifecycle.LifecycleAware;
 import org.jmxtrans.core.output.OutputWriter;
-import org.jmxtrans.core.query.embedded.InProcessServer;
-import org.jmxtrans.core.query.embedded.Query;
-import org.jmxtrans.core.query.embedded.ResultNameStrategy;
-import org.jmxtrans.core.query.embedded.Server;
+import org.jmxtrans.core.query.InProcessServer;
+import org.jmxtrans.core.query.Query;
+import org.jmxtrans.core.query.ResultNameStrategy;
+import org.jmxtrans.core.query.Server;
 import org.jmxtrans.core.results.QueryResult;
 import org.jmxtrans.utils.time.Clock;
 import org.jmxtrans.utils.time.Interval;
